Rename file according to convention
